ocr-cli/pyproject.toml
[project]
name = "ocr-cli"
version = "0.1.0"
description = "LLM-powered OCR for images and PDFs"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.50.0",
    "openai>=1.50.0",
    "pymupdf>=1.24.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "rich>=13.0.0",
]

[project.scripts]
ocr = "ocr_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ocr_cli"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    # PyMuPDF's SWIG-generated C extension emits these on import; not actionable.
    "ignore:builtin type.*has no __module__ attribute:DeprecationWarning",
]

